Timers can run in the background
For inspiration for this approach see 
https://github.com/duckduckgo/zeroclickinfo-goodies/blob/master/share/goodie/timer/timer.js

The issue was that if you went to a different tab from the timers, 
eventually they seemed to stop running. It looks like the browser either 
stops or slows down setTimeouts.

So instead, now a timed loop is running to govern all timers, and 
whenever that updates it passes the amount of time passed since the last 
update so the timers can remain accurate even if they are getting 
updated less frequently.